等权重融合全流程打通

+++ data/dataset/util.py
@@ -3,4 +3,4 @@ from sklearn.preprocessing import MinMaxScaler
 
 def preprocessing(x):
     model = MinMaxScaler()
-    return model.fit_transform(x)
+    return model.fit_transform(x.astype("float32"))
